Megacommit bcs I can't do things in order

+++ djlearn/todo/views.py
@@ -1,15 +1,30 @@
 from django.http import HttpResponseRedirect, HttpResponse
 from django.shortcuts import render, get_object_or_404
 from django.urls import reverse
-from .models import User, TodoItem, Team
+from accounts.views import SignUpView
+from.models import Todo_User, TodoItem, Team
 
 # Create your views here.
 
 def index(request):
-    return render(request, 'todo/index.html', {'user_list': User.objects.order_by('username')})
+    return render(request, 'todo/index.html', {'todo_user': Todo_User.objects.get(username=request.user.username) if request.user.is_authenticated else None})
 
 def todolist(request, pk):
-    user = get_object_or_404(User, pk=pk)
+    user = get_object_or_404(Todo_User, pk=pk)
+    perms = user.everyone_perms
+    if request.user.is_authenticated:
+        self = get_object_or_404(Todo_User, username=request.user.username)
+        if self == user:
+            perms = 'M'
+        elif self in user.friends.all():
+            match user.friend_perms:
+                case 'V':
+                    if perms != 'M':
+                        perms = 'V'
+                case 'M':
+                    perms = 'M'
+    if perms == 'P':
+        return HttpResponse("You do not have permission to view this user's todo list.", status=403)
     todo_items = user.todoitem_set.order_by('id')
     categories = Team.objects.filter(members=user).order_by('name')
     if categories.exists():
@@ -20,11 +35,19 @@ def todolist(request, pk):
                 categorized.append({'name': category.name, 'todo_items': items_in_category, 'modify': True})
         uncategorized_items = todo_items.filter(category__isnull=True)
         if uncategorized_items.exists():
-            categorized.append({'name': 'Uncategorized', 'todo_items': uncategorized_items, 'modify': True})
+            categorized.append({'name': 'Uncategorized', 'todo_items': uncategorized_items, 'modify': True if perms == 'M' else False})
         return render(request, 'todo/todolist.html', {'user': user, 'categories': categorized})
     else:
-        return render(request, 'todo/todolist.html', {'user': user, 'categories': [{'name': 'Uncategorized', 'todo_items': todo_items, 'modify': True}]})
-    
+        return render(request, 'todo/todolist.html', {'user': user, 'categories': [{'name': 'Uncategorized', 'todo_items': todo_items, 'modify': True if perms == 'M' else False}]})
+
+def check_modify_permission(user, self):
+    if self == user:
+        return True
+    elif user.everyone_perms == 'M':
+        return True
+    elif self in user.friends.all() and user.friend_perms == 'M':
+        return True
+    return False
 def update(request):
     # CSRF validation does not like any methods that aren't POST, so I just use that for everything
     if request.method == 'POST':
@@ -33,6 +56,8 @@ def update(request):
             try:
                 item_id = int(request.POST['item_id'])
                 item = get_object_or_404(TodoItem, pk=item_id)
+                if not check_modify_permission(item.user, get_object_or_404(Todo_User, username=request.user.username)):
+                    return HttpResponse("You do not have permission to modify this user's todo list.", status=403)
                 item.completed = not item.completed
                 item.save()
                 return HttpResponse("Item updated successfully", status=200)
@@ -42,6 +67,8 @@ def update(request):
             try:
                 item_id = int(request.POST['item_id'])
                 item = get_object_or_404(TodoItem, pk=item_id)
+                if not check_modify_permission(item.user, get_object_or_404(Todo_User, username=request.user.username) if request.user.is_authenticated else None):
+                    return HttpResponse("You do not have permission to modify this user's todo list.", status=403)
                 item.name = request.POST['name']
                 item.description = request.POST['description']
                 item.save()
@@ -51,10 +78,12 @@ def update(request):
         elif action == 'add':
             try:
                 user_id = int(request.POST['user_id'])
+                user = get_object_or_404(Todo_User, pk=user_id)
+                if not check_modify_permission(user, get_object_or_404(Todo_User, username=request.user.username) if request.user.is_authenticated else None):
+                    return HttpResponse("You do not have permission to modify this user's todo list.", status=403)
                 name = request.POST['name']
                 description = request.POST['description']
                 category_id = request.POST['category_id']
-                user = get_object_or_404(User, pk=user_id)
                 new_item = TodoItem(user=user, name=name, description=description, completed=False, assigned=False, category=Team.objects.get(pk=category_id) if category_id != '-1' else None)
                 new_item.save()
                 return HttpResponse("Item added successfully", status=201)
@@ -64,6 +93,8 @@ def update(request):
             try:
                 item_id = int(request.POST['item_id'])
                 item = get_object_or_404(TodoItem, pk=item_id)
+                if not check_modify_permission(item.user, get_object_or_404(Todo_User, username=request.user.username) if request.user.is_authenticated else None):
+                    return HttpResponse("You do not have permission to modify this user's todo list.", status=403)
                 item.delete()
                 return HttpResponse("Item deleted successfully", status=200)
             except (KeyError, ValueError):
@@ -71,7 +102,7 @@ def update(request):
         elif action == 'addUser':
             try:
                 username = request.POST['username']
-                new_user = User(username=username)
+                new_user = Todo_User(username=username)
                 new_user.save()
                 return HttpResponse("User created successfully", status=201)
             except (KeyError, ValueError):
@@ -79,17 +110,16 @@ def update(request):
         elif action == 'deleteUser':
             try:
                 user_id = int(request.POST['user_id'])
-                user = get_object_or_404(User, pk=user_id)
+                user = get_object_or_404(Todo_User, pk=user_id)
                 user.delete()
                 return HttpResponseRedirect(reverse('todo:index'))
             except (KeyError, ValueError):
                 return HttpResponse("Invalid user ID", status=400)
         elif action == 'addFriend':
             try:
-                user_id = int(request.POST['user_id'])
-                user = get_object_or_404(User, pk=user_id)
+                user = get_object_or_404(Todo_User, username=request.user.username)
                 friend_name = request.POST['friend_id']
-                friend = get_object_or_404(User, username=friend_name)
+                friend = get_object_or_404(Todo_User, username=friend_name)
                 if friend == user:
                     return HttpResponse("Cannot add yourself as a friend", status=400)
                 if friend in user.friends.all():
@@ -103,10 +133,9 @@ def update(request):
                 return HttpResponse("Invalid data provided", status=400)
         elif action == 'acceptFriend':
             try:
-                user_id = int(request.POST['user_id'])
-                user = get_object_or_404(User, pk=user_id)
+                user = get_object_or_404(Todo_User, username=request.user.username)
                 friend_id = int(request.POST['friend_id'])
-                friend = get_object_or_404(User, pk=friend_id)
+                friend = get_object_or_404(Todo_User, pk=friend_id)
                 if friend not in user.friend_requests.all():
                     return HttpResponse("No friend request from this user", status=400)
                 user.friends.add(friend)
@@ -117,10 +146,9 @@ def update(request):
                 return HttpResponse("Invalid data provided", status=400)
         elif action == 'removeFriend':
             try:
-                user_id = int(request.POST['user_id'])
-                user = get_object_or_404(User, pk=user_id)
+                user = get_object_or_404(Todo_User, username=request.user.username)
                 friend_id = int(request.POST['friend_id'])
-                friend = get_object_or_404(User, pk=friend_id)
+                friend = get_object_or_404(Todo_User, pk=friend_id)
                 if friend not in user.friends.all():
                     return HttpResponse("Not friends with this user", status=400)
                 user.friends.remove(friend)
@@ -130,10 +158,9 @@ def update(request):
                 return HttpResponse("Invalid data provided", status=400)
         elif action == 'declineFriend':
             try:
-                user_id = int(request.POST['user_id'])
-                user = get_object_or_404(User, pk=user_id)
+                user = get_object_or_404(Todo_User, username=request.user.username)
                 friend_id = int(request.POST['friend_id'])
-                friend = get_object_or_404(User, pk=friend_id)
+                friend = get_object_or_404(Todo_User, pk=friend_id)
                 if friend not in user.friend_requests.all():
                     return HttpResponse("No friend request from this user", status=400)
                 user.friend_requests.remove(friend)
@@ -143,10 +170,9 @@ def update(request):
                 return HttpResponse("Invalid data provided", status=400)
         elif action == 'cancelFriendRequest':
             try:
-                user_id = int(request.POST['user_id'])
-                user = get_object_or_404(User, pk=user_id)
+                user = get_object_or_404(Todo_User, username=request.user.username)
                 friend_id = int(request.POST['friend_id'])
-                friend = get_object_or_404(User, pk=friend_id)
+                friend = get_object_or_404(Todo_User, pk=friend_id)
                 if user not in friend.friend_requests.all():
                     return HttpResponse("No sent friend request to this user", status=400)
                 friend.friend_requests.remove(user)
@@ -154,10 +180,9 @@ def update(request):
                 return HttpResponse("Friend request canceled", status=200)
             except (KeyError, ValueError):
                 return HttpResponse("Invalid data provided", status=400)
-        elif action == 'setFriendPerms':
+        elif action == 'updateFriendPerms':
             try:
-                user_id = int(request.POST['user_id'])
-                user = get_object_or_404(User, pk=user_id)
+                user = get_object_or_404(Todo_User, username=request.user.username)
                 perms = request.POST['perms']
                 if perms not in ['P', 'V', 'M']:
                     return HttpResponse("Invalid permissions value", status=400)
@@ -166,13 +191,39 @@ def update(request):
                 return HttpResponse("Friend permissions updated", status=200)
             except (KeyError, ValueError):
                 return HttpResponse("Invalid data provided", status=400)
+        elif action == 'updateEveryonePerms':
+            try:
+                user = get_object_or_404(Todo_User, username=request.user.username)
+                perms = request.POST['perms']
+                if perms not in ['P', 'V', 'M']:
+                    return HttpResponse("Invalid permissions value", status=400)
+                user.everyone_perms = perms
+                user.save()
+                return HttpResponse("Everyone permissions updated", status=200)
+            except (KeyError, ValueError):
+                return HttpResponse("Invalid data provided", status=400)
+        elif action == 'setEmail':
+            try:
+                request.user.email = request.POST['email']
+                request.user.save()
+                return HttpResponse("Email updated successfully", status=200)
+            except (KeyError, ValueError):
+                return HttpResponse("Invalid data provided", status=400)
+        elif action == 'setDisplayName':
+            try:
+                user = get_object_or_404(Todo_User, username=request.user.username)
+                user.display_name = request.POST['display_name']
+                user.save()
+                return HttpResponse("Display name updated successfully", status=200)
+            except (KeyError, ValueError):
+                return HttpResponse("Invalid data provided", status=400)
         else:
             return HttpResponse("Invalid action", status=400)
     else:
         return HttpResponse("Method Not Allowed: " + request.method, status=405)
 
 def edit(request, pk):
-    user = get_object_or_404(User, pk=pk)
+    user = get_object_or_404(Todo_User, pk=pk)
     todo_items = TodoItem.objects.filter(user=user)
     categories = Team.objects.filter(members=user).order_by('name')
     if categories.exists():
@@ -189,20 +240,20 @@ def edit(request, pk):
         return render(request, 'todo/edit.html', {'user': user, 'categories': [{'name': 'Uncategorized', 'todo_items': todo_items, 'modify': True, 'id': -1, 'super': False}]})
     
 def user(request, pk):
-    user = get_object_or_404(User, pk=pk)
+    user = get_object_or_404(Todo_User, pk=pk)
     categories = Team.objects.filter(members=user)
-    friend_requests = User.objects.filter(friend_requests=user)
+    friend_requests = Todo_User.objects.filter(friend_requests=user)
     incoming_requests = user.friend_requests.all()
     friends = user.friends.all()
-    self = get_object_or_404(User, username=request.user.username) if request.user.is_authenticated else None
+    self = get_object_or_404(Todo_User, username=request.user.username) if request.user.is_authenticated else None
     return render(request, 'todo/user.html', {'user': user, 'categories': categories, "friend_requests": friend_requests, "incoming_requests": incoming_requests, "friends": friends, "self": self})
 
 def user_redirect(request):
     if request.user.is_authenticated:
         try:
-            user = User.objects.get(username=request.user.username)
+            user = Todo_User.objects.get(username=request.user.username)
             return HttpResponseRedirect(reverse('todo:user', args=(user.id,)))
-        except User.DoesNotExist:
+        except Todo_User.DoesNotExist:
             return HttpResponse("User not found. Try logging in again.", status=404)
     else:
         return HttpResponseRedirect(reverse('login'))
@@ -213,3 +264,12 @@ def team(request, pk):
     leaders = team.leaders.all()
     todo_items = TodoItem.objects.filter(category=team)
     return render(request, 'todo/team.html', {'team': team, 'members': members, 'leaders': leaders, 'todo_items': todo_items})
+
+def signup(request):
+    if request.method == 'POST':
+        signup = SignUpView.as_view()(request)
+        if signup:
+            Todo_User.objects.create(username=request.POST['username'], display_name=request.POST['username'])
+            return signup
+    else:
+        return HttpResponse(status=405)
